requirements.txt
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.0.0
PyPDF2==3.0.1
argparse==1.4.0
tqdm==4.65.0
reportlab==4.0.4 